+++ index.html
@@ -230,7 +230,5 @@
     </div>
   </template>
 <script src="js/main.js" type="module"></script>
-<script src="js/util.js" type="module"></script>
-<script src="js/data.js" type="module"></script>
 </body>
 </html>
